Keep connectivity after installing FRR on debian
When installing the FRR package, something happens to cause the default
routes to go away. In order to fix this, we simply need to run `netplan
apply`. But this needs to be chained off of the package install. So the
slightly goofy `shell` module for the package install and `netplan
apply` is a tad hacky, but provides a specific pattern we're attempting
to hit.

Signed-off-by: David Wahlstrom <david.wahlstrom@dreamhost.com>